@fulmicoton 2023-11-13
Can you try an fst table? (It should faster... build times should be impacted however.)